5.1. разбивает код на модули
